+++ packages/core/core/prompt/index.ts
@@ -1,4 +1,5 @@
 export * from './content-part';
 export * from './convert-to-core-messages';
 export * from './data-content';
+export * from './invalid-message-role-error';
 export * from './message';