+++ packages/core/tests/e2e/next-server/app/streamable/page.js
@@ -1,0 +1,20 @@
+import {
+  streamableUI,
+  streamableValue,
+  streamableUIAppend,
+  streamableUIError,
+} from './actions';
+import { Client } from './client';
+
+export default function Page() {
+  return (
+    <Client
+      actions={{
+        streamableUI,
+        streamableValue,
+        streamableUIAppend,
+        streamableUIError,
+      }}
+    />
+  );
+}