+++ packages/provider-utils/src/post-to-api.ts
@@ -3,6 +3,9 @@ import { extractResponseHeaders } from './extract-response-headers';
 import { isAbortError } from './is-abort-error';
 import { ResponseHandler } from './response-handler';
 
+// use function to allow for mocking in tests:
+const getOriginalFetch = () => fetch;
+
 export const postJsonToApi = async <T>({
   url,
   headers,
@@ -10,6 +13,7 @@ export const postJsonToApi = async <T>({
   failedResponseHandler,
   successfulResponseHandler,
   abortSignal,
+  fetch,
 }: {
   url: string;
   headers?: Record<string, string | undefined>;
@@ -17,6 +21,7 @@ export const postJsonToApi = async <T>({
   failedResponseHandler: ResponseHandler<APICallError>;
   successfulResponseHandler: ResponseHandler<T>;
   abortSignal?: AbortSignal;
+  fetch?: ReturnType<typeof getOriginalFetch>;
 }) =>
   postToApi({
     url,
@@ -31,6 +36,7 @@ export const postJsonToApi = async <T>({
     failedResponseHandler,
     successfulResponseHandler,
     abortSignal,
+    fetch,
   });
 
 export const postToApi = async <T>({
@@ -40,6 +46,7 @@ export const postToApi = async <T>({
   successfulResponseHandler,
   failedResponseHandler,
   abortSignal,
+  fetch = getOriginalFetch(),
 }: {
   url: string;
   headers?: Record<string, string | undefined>;
@@ -50,6 +57,7 @@ export const postToApi = async <T>({
   failedResponseHandler: ResponseHandler<Error>;
   successfulResponseHandler: ResponseHandler<T>;
   abortSignal?: AbortSignal;
+  fetch?: ReturnType<typeof getOriginalFetch>;
 }) => {
   try {
     // remove undefined headers: